consumer/requirements/requirements.txt
kafka-python==2.0.2
beautifultable==1.0.1
psycopg2-binary==2.8.6
